nlp/requirements.txt
accelerate>=0.30,<2.0
datasets>=2.19,<4.0
evaluate>=0.4,<1.0
Flask>=3.0,<4.0
huggingface-hub>=0.30,<1.0
neo4j>=5.20,<6.0
Pillow>=10.4,<12.0
pydantic>=2.7,<3.0
pydantic-settings>=2.3,<3.0
pytesseract>=0.3.10,<0.4.0
pypdf>=4.2,<6.0
pytest>=8.2,<9.0
reportlab>=4.2,<5.0
requests>=2.32,<3.0
scikit-learn>=1.5,<2.0
seqeval>=1.2,<2.0
spacy>=3.7,<4.0
torch>=2.3,<3.0
transformers>=4.41,<5.0
xgboost>=2.0,<3.0
